CanvasSynch/pyproject.toml
[tool.poetry]
name = "canvas-ruz"
version = "0.1.0"
description = ""
authors = ["KnowYourselves <rsalvarez@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
